% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTSV.R
\name{SVGene}
\alias{SVGene}
\title{Report spatially variable genes}
\usage{
SVGene(Q_val, thre.alpha)
}
\arguments{
\item{Q_val}{A G by 2K q-value matrix, where G is the number of genes and K is the number of cell types.}

\item{thre.alpha}{numeric, a q-value threshold to control FDR less than thre.alpha.}

}
\value{
A list with a G by 2K 0-1 matrix and a list with names of SV genes in each cell type. The first K columns of the 0-1 matrix correspond to the coordinate of \eqn{S_1}, and the last K columns to the coordinate of \eqn{S_2}.
\item{SV}{A G by 2K 0-1 matrix. The first K columns correspond to the coordinate of \eqn{S_1}, the last K columns to the coordinate of \eqn{S_2}.}
\item{SVGene}{A list with names of SV genes in each cell type.}
}
\description{
Report spatially variable genes using the results from the function ctsv.
}
