% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTSV.R
\name{ctsv}
\alias{ctsv}
\title{Detection of cell-type-specific spatially variable genes}
\usage{
ctsv(Y, loc, W, num_core = 1)
}
\arguments{
\item{Y}{An n by G bulk ST raw count data matrix, where n is the spot number and G is the gene number. Each row stands for a spot and each column represents a gene. The column names of Y (i.e. gene names) must be specified, and the row names of Y must be consistent with row names of loc and W.}

\item{loc}{An n by 2 location matrix, where each row corresponds to a spot and records the two-dimensional coordinate of that spot.rd}

\item{W}{An n by K cell-type proportion matrix, where K is the number of cell types. The column names of W are cell type names.}

\item{num_core}{Number of cores if using paralleling. The default is one.}
}
\value{
A list with a G by 2K matrix of p-values and a G by 2K matrix of q-values.
\item{pval}{A G by 2K matrix of p-values. The first K columns correspond to the first coordinate, and the last K columns to the second coordinate.}
\item{qval}{A G by 2K matrix of q-values. The first K columns correspond to the first coordinate, and the last K columns to the second coordinate.}
}
\description{
Detection of cell-type-specific spatially variable genes
}
\examples{
library(CTSV)
#read example data
data(CTSVexample_data)
Y <- CTSVexample_data[[4]]
W <- CTSVexample_data[[3]]
loc <- CTSVexample_data[[2]]
gamma_true <- CTSVexample_data[[1]]
# gene number
G <- ncol(Y)
# spot number
n <- nrow(Y)
# cell type number
K <- ncol(W)

# SV genes in each cell type:
print(colnames(Y)[which(gamma_true[,1] == 1)])
print(colnames(Y)[which(gamma_true[,2] == 1)])
# Number of SV genes at the aggregated level:
print(sum(rowSums(gamma_true)>0))

#--- Run CTSV ----
result <- ctsv(Y,loc,W,num_core = 8)
# View on q-value matrix
head(result$qval)
# detect SV genes
re <- svGene(result$qval,0.05)
#SV genes in each cell type:
print(re$SVGene)
}
